Refactored avatar mqueueing
